Split string_to_rcst into multiple files
